fix(refactor): fixed test and improve code structure
